server/requirements.txt
requests
python-dotenv
wget
fastapi
uvicorn